requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.0.1
django-admin-interface==0.17.3
django-colorfield==0.4.5
django-easy-pdf==0.1.1
django-extensions==3.1.3
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-material==1.9.0
django-utils-six==2.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
Pillow==8.4.0
psycopg2==2.8.6
pydot==1.4.2
pyparsing==3.0.3
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.2
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
webencodings==0.5.1
xhtml2pdf==0.2.5
